% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditions.R
\docType{data}
\name{conditions}
\alias{conditions}
\title{Conditions for all dyes screens}
\format{
get to this later
}
\usage{
conditions
}
\description{
A dataset containing conditions for all dye screens
}
\keyword{datasets}
