% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\docType{data}
\name{heatmap}
\alias{heatmap}
\title{Assignments for all dyes screens}
\format{
A data frame with 22,223 rows rows and 19 variables:
}
\usage{
heatmap
}
\description{
A dataset containing all proteins and dyes screened, and the resulting assignment. Also includes screening conditions.
}
\keyword{datasets}
