% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_ggproto.R
\docType{data}
\name{heatmap_ggproto}
\alias{heatmap_ggproto}
\title{A ggproto object of the heatmap plot}
\format{
get to this later
}
\usage{
heatmap_ggproto
}
\description{
Ready to display
}
\keyword{datasets}
