% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot_raw.R
\name{plot_raw}
\alias{plot_raw}
\title{plot_raw}
\usage{
plot_raw(
  tidied_screen,
  .temp_col = "Temperature",
  .value_col = "value",
  .color_col = "channel_f",
  .linetype_col = "type",
  .variable_col = "var_num",
  .line_size = 0.3,
  .line_alpha = 0.8,
  .color_legend_title = "Channel",
  .linetype_legend_title = "",
  .scale_color_vals = c(Cy5.5 = "#67000d", Cy5 = "#a50f15", ROX = "#ef3b2c", TAMRA =
    "#f16913", JOE = "#74c476", FAM = "#2171b5"),
  .scale_linetype_vals = c(buffer = "dashed", protein = "solid"),
  .use_default_x_scaling = TRUE,
  .x_scale_breaks = NULL,
  ...
)
}
\value{
A ggproto object containing a plot of the raw data
}
\description{
A funcion to plot raw screen results
}
